refactor(policy): split document adapters
