+++ dashboard/src/settings/tray-settings-panel.tsx
@@ -13,6 +13,16 @@ interface TrayActionState {
 
 const IDLE_STATE: TrayActionState = { status: "idle", message: "" };
 
+function actionFeedbackClassName(status: TrayActionStatus): string {
+  if (status === "error") {
+    return "rounded-lg border border-red-200 bg-red-50 p-3 text-sm text-red-700";
+  }
+  if (status === "success") {
+    return "rounded-lg border border-green-200 bg-green-50 p-3 text-sm text-green-700";
+  }
+  return "rounded-lg border border-slate-200 bg-slate-50 p-3 text-sm text-slate-600";
+}
+
 export function TraySettingsPanel(): JSX.Element {
   const [trayStatus, setTrayStatus] = useState<TrayStatusPayload | null>(null);
   const [statusLoading, setStatusLoading] = useState(true);
@@ -119,13 +129,7 @@ export function TraySettingsPanel(): JSX.Element {
           {/* Action feedback */}
           {actionState.status !== "idle" && (
             <div
-              className={
-                actionState.status === "error"
-                  ? "rounded-lg border border-red-200 bg-red-50 p-3 text-sm text-red-700"
-                  : actionState.status === "success"
-                    ? "rounded-lg border border-green-200 bg-green-50 p-3 text-sm text-green-700"
-                    : "rounded-lg border border-slate-200 bg-slate-50 p-3 text-sm text-slate-600"
-              }
+              className={actionFeedbackClassName(actionState.status)}
               role={actionState.status === "error" ? "alert" : "status"}
             >
               {actionState.status === "loading" ? "Working…" : actionState.message}
